Split commands out to individual files
